\name{AnnotateExtra-methods}
\docType{methods}
\alias{AnnotateExtra-methods}
\alias{AnnotateExtra,ExomeDepth-method}
\title{ Additional annotations for ExomeDepth objects}
\description{
  Takes a GRanges object, typically a list of reference calls, and add
  these annotations to a ExomeDepth object.
}
\section{Methods}{
\describe{

\item{\code{signature(x = "ExomeDepth")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
