% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countBamInGranges.R
\docType{data}
\name{Conrad.hg19.common.CNVs}
\alias{Conrad.hg19.common.CNVs}
\title{Conrad et al common CNVs}
\format{A data frame with common CNV calls.}
\source{
Conrad et al, Origins and functional impact of copy number variation in the human genome, Nature 2010
}
\description{
Positions of common CNV calls (detected in a panel of 42 sample) from the
Conrad et al paper (Nature 2010). This is build hg19 of the human genome.
}
\keyword{datasets}
