% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countBamInGranges.R
\docType{data}
\name{ExomeCount}
\alias{ExomeCount}
\title{Example dataset for ExomeDepth}
\format{
A data frame with 25592 observations on the following 9 variables:
\itemize{
\item chromosome, Character vector with chromosome names (only chromosome 1 in that case)
\item start, start of exons
\item end, end of exons
\item exons, character name of exons
\item camfid.032KA_sorted_unique.bam
\item camfid.033ahw_sorted_unique.bam
\item camfid.034pc_sorted_unique.bam
\item camfid.035if_sorted_unique.bam
\item GC, a numeric vector with the GC content
}
}
\source{
Dataset generated in collaboration with Sergey Nejentsev, University of Cambridge.
}
\description{
An example dataset of 4 exome samples, chromosome 1 only.
}
\keyword{datasets}
