\name{exons.hg19}
\alias{exons.hg19}
\docType{data}
\title{
  Positions of exons on build hg19 of the human genome
}
\description{
  Exon position extracted from the ensembl database version 71.
}
\usage{data(exons.hg19)}
\format{
  A data frame with 192,379 observations on the following 4 variables.
  \describe{
    \item{\code{chromosome}}{a factor with levels \code{1} \code{10} \code{11} \code{12} \code{13} \code{14} \code{15} \code{16} \code{17} \code{18} \code{19} \code{2} \code{20} \code{21} \code{22} \code{3} \code{4} \code{5} \code{6} \code{7} \code{8} \code{9}}
    \item{\code{start}}{a numeric vector}
    \item{\code{end}}{a numeric vector}
    \item{\code{name}}{A character vector of names for the exon(s)}
  }
}
\source{
  Ensemble database version 71.
}
\keyword{datasets}
