% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{qbetabinom.ab}
\alias{qbetabinom.ab}
\title{Quantile function for the beta-binomial distribution}
\usage{
qbetabinom.ab(p, size, shape1, shape2)
}
\arguments{
\item{p}{Mean value of the beta-binomial distribution.}

\item{size}{Size of the beta-binomial.}

\item{shape1}{First parameter of the beta distribution for p.}

\item{shape2}{Second parameter of the beta distribution for p.}
}
\value{
A quantile of the distribution.
}
\description{
Standard qbetabinomial.ab function which is missing from the VGAM package.
}
\examples{
qbetabinom.ab(p = 0.5, size = 50, shape1 = 0.2, shape2 = 0.25)
}
\seealso{
\code{VGAM} package.
}
