% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{qbetabinom}
\alias{qbetabinom}
\title{Quantile for betabin function}
\usage{
qbetabinom(p, size, phi, prob)
}
\arguments{
\item{p}{Point of the distribution from which one is looking for the
quantile}

\item{size}{Sample size of the random variable}

\item{phi}{Over-dispersion parameter}

\item{prob}{Mean probability of the binomial distribution}
}
\value{
A real number corresponding to the quantile p.
}
\description{
Quantile function for the betabinomal distribution using the p/phi
parameterisation.
}
\details{
Filling a gap in the VGAM package.
}
\examples{
qbetabinom(p = 0.2, size = 50, phi = 0.4, prob = 0.3)
qbetabinom(p = 0.2, size = 50, phi = 0.1, prob = 0.8)
}
\seealso{
VGAM R package.
}
\author{
Vincent Plagnol
}
