% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definition.R
\name{TestCNV,ExomeDepth-method}
\alias{TestCNV,ExomeDepth-method}
\alias{TestCNV}
\title{TestCNV}
\usage{
\S4method{TestCNV}{ExomeDepth}(x, chromosome, start, end, type)
}
\arguments{
\item{x}{ExomeDepth object}

\item{chromosome}{Character, chromosome name.}

\item{start}{Numeric, start of the tested CNV}

\item{end}{Numeric, end of the tested CNV}

\item{type}{Character, must be either \code{deletion} or \code{duplication}.}
}
\description{
Computes the Bayes Factor in favour of a CNV defined by position and type.
}
